requirements.txt
Pillow
requests
pyyaml